CheckPoints and More Folders

+++ Assets/Scripts/Game/Weapons/Bullets/BulletData.cs
@@ -5,7 +5,7 @@ namespace Game.Weapons.Bullets
     [CreateAssetMenu(menuName = "Game/Weapons/BulletData")]
     public class BulletData : ScriptableObject
     {
-        public float Damage => _damage;
+        public int Damage => _damage;
         public float Speed => _speed;
         public float MaxDistance => _maxDistance;
         
@@ -13,7 +13,7 @@ namespace Game.Weapons.Bullets
         [SerializeField] private Bullet _prefab;
         
         [Header("Parameters")]
-        [SerializeField] private float _damage;
+        [SerializeField] private int _damage;
         [SerializeField] private float _speed;
         [SerializeField] private int _maxDistance;
